requirement.txt
# Core Dependencies
numpy==1.24.4
pandas==2.1.1
Pillow==10.0.1
opencv-python==4.8.0.76
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
scikit-image==0.24.0
lpips==0.1.4

# CLIP (from OpenAI's official repository)
git+https://github.com/openai/CLIP.git

# Optional (for progress visualization and plotting)
tqdm==4.66.1
matplotlib==3.8.0
